refactor: WebMvcConfigurer 구성 통합
* CORS 구성과 WebMvc 구성 두 개로 나누어져 있던 것을 하나로 합침
- CorsConfiguration을 제거하고 WebMvcConfiguration으로 옮김
