+++ core/core-api/src/main/java/com/woopaca/taximate/core/api/common/config/WebMvcConfiguration.java
@@ -4,6 +4,7 @@ import com.woopaca.taximate.core.api.auth.interceptor.AuthenticateInterceptor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.client.RestClient;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
@@ -24,5 +25,13 @@ public class WebMvcConfiguration implements WebMvcConfigurer {
     @Bean
     public RestClient restClient() {
         return RestClient.create();
+    }
+
+    @Override
+    public void addCorsMappings(CorsRegistry registry) {
+        registry.addMapping("/**")
+                .allowedOrigins("http://localhost:5173", "https://localhost:5173")
+                .allowedMethods("GET", "POST", "PUT", "DELETE")
+                .allowCredentials(true);
     }
 }
